+++ packages/demo/src/js/app.ts
@@ -1,9 +1,6 @@
 import { Base, createApp } from '@studiometa/js-toolkit';
-import { Cursor } from '@studiometa/ui';
-import { createApp as createVueApp } from 'vue';
-import VueComponent from './VueComponent.vue';
-import config from '#js/config.yaml';
-import configRaw from './config.yaml?raw';
+
+const demoToReversedValues = [1, 2, 3].toReversed();
 
 /**
  * App class.
@@ -16,54 +13,10 @@ class App extends Base {
   static config = {
     name: 'App',
     log: true,
-    refs: ['content', 'vue'],
-    ...config,
-    components: {
-      Cursor,
-      // eslint-disable-next-line import/extensions
-      Component: () => import(/* webpackPrefetch: true */ './components/Component.js'),
-    },
   };
 
-  vue;
-
-  VueComponent;
-
-  /**
-   * Mounted hook.
-   */
-  async mounted() {
-    this.$log('config', config, configRaw, this.$options);
-    this.content = 'mounted';
-    this.VueComponent = VueComponent;
-  }
-
-  /**
-   * Load the Vue app on click.
-   */
-  async onClick() {
-    if (this.vue) {
-      return;
-    }
-
-    const { default: VueCounter } = await import(/* webpackPreload: true */ './Counter.vue');
-    this.vue = createVueApp(VueCounter);
-    this.vue.mount(this.$refs.vue);
-  }
-
-  /**
-   * Resized hook.
-   */
-  resized() {
-    this.content = 'resized';
-  }
-
-  /**
-   * Set the component's content.
-   * @param {string} value The content to add.
-   */
-  set content(value) {
-    this.$refs.content.innerHTML += `<br>${value}`;
+  mounted() {
+    this.$log(demoToReversedValues);
   }
 }